# tickets/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events",
                "description": "Returns paginated list of events with optional filters",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Start date (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "End date (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "start_at",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events"
                    },
                    "404": {
                        "description": "No events found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create new event",
                "description": "Admin only - Creates a new event with seats",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "venue",
                                    "start_at"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Summer Concert"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "summer-concert-2026"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Amazing summer concert"
                                    },
                                    "venue": {
                                        "type": "string",
                                        "example": "Arena Belgrade"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Belgrade"
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-07-15 20:00:00"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-07-15 23:00:00"
                                    },
                                    "rows": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "columns": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully"
                    },
                    "403": {
                        "description": "Only admins can create events"
                    }
                }
            }
        },
        "/api/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get single event",
                "description": "Returns event details with ticket types",
                "operationId": "2ee679ce749d46144599318b5ba161f5",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "description": "Admin only - Updates event details",
                "operationId": "2e0419cd91b178e7bcffa79a3ebc44d6",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "venue": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully"
                    },
                    "403": {
                        "description": "Only admins can update events"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete event",
                "description": "Admin only - Deletes an event",
                "operationId": "f4dd0f8b74a8e726ca902d975bb372d8",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully"
                    },
                    "403": {
                        "description": "Only admins can delete events"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
...
    },
...
}